Need Help with Activities Total
I'm trying to get the div with id of "cost" to display the "Your total cost is : $(totalCost)." (totalCost) being my let variable that holds the running total. The text on the form actually displays the (totalCost) and not the amount held in that variable. Once I figure how to fix this, I'm planning to do if, else functions for each checked input. I also can't seem to make totalCost += 100 sum into the totalCost variable.